Restaurants/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
pillow==11.0.0
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
